Fix #352: Permission denied when creating a folder
When SFTP does not have permission to list a folder, but does have access to subsequent folders, which is very often the case (e.g. mkdir /home/<user>/mynewfolder but no permission to ls /home), creating the folder fails. This change keeps trying to create a the folder in the next folder even if it doesn't have access to the current folder.